feat: refactor evm into modules and move tests into pytest directory
