+++ src/zkevm_specs/__init__.py
@@ -1,2 +1,4 @@
 from . import encoding
+from . import evm
 from . import opcode
+from . import util
