+++ src/zkevm_specs/evm/__init__.py
@@ -1,0 +1,5 @@
+from .execution_result import *
+from .main import *
+from .opcode import *
+from .step import *
+from .table import *
