+++ src/zkevm_specs/evm/execution_result/error_stack_overflow.py
@@ -1,0 +1,9 @@
+from ..opcode import Opcode
+from ..step import Step
+from ..table import FixedTableTag
+
+
+def error_stack_overflow(curr: Step, next: Step, r: int, opcode: Opcode):
+    curr.fixed_lookup(FixedTableTag.StackOverflow, [opcode, curr.call_state.stack_pointer])
+
+    # TODO: Return to caller's state or go to next tx
